Added plugin API for incremental parser
Also moved indentation extension into a plugin
