+++ lib/eco/test/test_eco.py
@@ -1089,6 +1089,15 @@ def z():
         self.treemanager.key_normal("\"")
         assert self.parser.last_status == True
 
+    def test_indentation_comment(self):
+        self.reset()
+        inputstring = """class X:
+    # test
+    pass"""
+        self.treemanager.import_file(inputstring)
+        assert self.parser.last_status == True
+        
+
 class Test_NestedLboxWithIndentation():
     def setup_class(cls):
         parser, lexer = calc.load()
